style updates; avoid composes
