+++ src/components/chatbox/message.module.css
@@ -1,0 +1,94 @@
+.message {
+	display: flex;
+	flex-direction: row;
+	gap: 12px;
+	padding: 20px 16px;
+	background: var(--token-color-palette-neutral-0);
+}
+
+.message_user {
+	border-bottom: 1px solid var(--token-color-palette-neutral-300);
+}
+
+.message_avatar,
+.message_avatar img {
+	width: 40px;
+	height: 40px;
+	border-radius: 6px;
+}
+
+.message_content {
+	padding-top: 10px; /* simulate vertically centering the first line with avatar/icon */
+	display: flex;
+	flex: 1;
+	flex-direction: column;
+}
+
+.message_user_input {
+	color: var(--token-color-foreground-primary, #3b3d45);
+	white-space: pre-line; /* visually respect user-inputted line breaks */
+}
+
+.message_gutter {
+	min-width: 40px;
+}
+
+.message_assistant {
+	background: var(--token-color-surface-strong); /* #f1f2f3 */
+}
+
+.message_icon {
+	--icon-color: var(--token-color-foreground-highlight-on-surface, #911ced);
+}
+
+.message_markdown {
+	color: var(--token-color-foreground-primary, #3b3d45);
+	& pre {
+		overflow-x: auto;
+		white-space: pre-wrap;
+		word-wrap: break-word;
+	}
+}
+
+.message_AssistantMessageFooter {
+	width: 100%;
+	padding: 24px 0 0;
+}
+
+.message_AssistantMessageFooterHidden {
+	display: none;
+}
+
+.message_divider {
+	display: block;
+	width: 100%;
+	border-bottom: 1px solid var(--token-color-palette-neutral-300);
+	margin-bottom: 24px;
+}
+
+.message_actionButtons {
+	display: flex;
+	flex-direction: row;
+	gap: 16px;
+}
+
+.rating_like {
+	border: 1px solid var(--token-color-border-success) !important;
+	background: var(--token-color-surface-success) !important;
+	color: var(--token-color-foreground-success) !important;
+}
+.rating_dislike {
+	border: 1px solid var(--token-color-border-critical) !important;
+	background: var(--token-color-surface-critical) !important;
+	color: var(--token-color-foreground-critical) !important;
+}
+
+.message_feedbackForm {
+	&.message_feedbackFormVisible {
+		margin-top: 30px;
+	}
+}
+
+.message_icon_error {
+	--icon-color: var(--token-color-foreground-critical);
+}
